fix(admin): stop circular re-export on supabase-config route
Admin build failed with Maximum call stack size exceeded because
@/app/api/... resolved to the admin route re-exporting itself.

Move handler to lib/api/public/supabase-config.ts and import from both apps.
