+++ app/api/public/supabase-config/route.ts
@@ -1,25 +1,9 @@
 // PUBLIC ROUTE: Supabase anon key is intended for browser use (RLS-protected).
 
-import { NextResponse } from 'next/server';
-import { getServerPublicSupabaseConfig } from '@/lib/supabase/public-config';
+import { getSupabasePublicConfigResponse } from '@/lib/api/public/supabase-config';
 
 export const dynamic = 'force-dynamic';
 
 export async function GET() {
-  const config = getServerPublicSupabaseConfig();
-  if (!config) {
-    return NextResponse.json(
-      { error: 'Authentication service is not configured on this deployment.' },
-      { status: 503 },
-    );
-  }
-
-  return NextResponse.json(
-    { url: config.url, anonKey: config.anonKey },
-    {
-      headers: {
-        'Cache-Control': 'private, no-store',
-      },
-    },
-  );
+  return getSupabasePublicConfigResponse();
 }
